xresources: Set urxvt geometry to avoid being in middle of screen
Since urxvt 9.31 the following code will end up with the cursor in the
middle of the screen after prompt drawing:

+          // wing, instead of wrap
+          screen.cur.row += nrow - prev_nrow;

This means if the geometry is smaller than the eventual resize from the
window manager, the prompt is left in the middle of the screen. Make it
big to avoid that.